Move the facility contexts
