+++ src/components/rt-timeseries/RtTimeseriesContainer.tsx
@@ -2,13 +2,13 @@ import hexAlpha from "hex-alpha";
 import React, { useContext, useState } from "react";
 import styled from "styled-components";
 
+import { FacilityContext } from "../../contexts/facility-context/FacilityContext";
 import Colors from "../../design-system/Colors";
 import HelpButtonWithTooltip from "../../design-system/HelpButtonWithTooltip";
 import Loading from "../../design-system/Loading";
 import { RtData } from "../../infection-model/rt";
 import { updateFacilityRtData } from "../../infection-model/rt";
 import AddCasesModal from "../../page-multi-facility/AddCasesModal";
-import { FacilityContext } from "../../page-multi-facility/FacilityContext";
 import { Facility } from "../../page-multi-facility/types";
 import RtTimeseries from "./RtTimeseries";
