+++ src/page-multi-facility/FacilityInputForm.tsx
@@ -3,6 +3,7 @@ import React, { useContext, useState } from "react";
 import styled from "styled-components";
 
 import RtTimeseries from "../components/rt-timeseries";
+import { FacilityContext } from "../contexts/facility-context/FacilityContext";
 import { deleteFacility, saveFacility } from "../database/index";
 import Colors from "../design-system/Colors";
 import iconDuplicatePath from "../design-system/icons/ic_duplicate.svg";
@@ -19,7 +20,6 @@ import MitigationInformation from "../impact-dashboard/MitigationInformation";
 import useModel from "../impact-dashboard/useModel";
 import { updateFacilityRtData } from "../infection-model/rt";
 import AddCasesModal from "./AddCasesModal";
-import { FacilityContext } from "./FacilityContext";
 import FacilityProjections from "./FacilityProjections";
 import HistoricalCasesChart from "./HistoricalCasesChart";
 import LocaleInformationSection from "./LocaleInformationSection";
